Add config for hosts override file
The main use of this is for dns based blackholeing of ad sites a la
github.com/StevenBlack/hosts. You are probably better off just using
the /etc/hosts file instead of this.

My main motivation for this is to avoid having bash do hostname
completion on my block list.